+++ src/z80/mod.rs
@@ -1,12 +1,12 @@
 //! A back-end targeting the Z80, a well-known 8-bit retro CPU.
-mod calling_conventions;
+mod sdcccall1;
 mod dataflow;
 mod diss;
 mod emu;
 mod isa;
 mod subroutine;
 
-pub use calling_conventions::SdccCall1;
+pub use sdcccall1::SdccCall1;
 pub use emu::Emulator;
 pub use isa::Insn;
 pub use subroutine::Subroutine;
